Added problem 15 Tidied up some other code

+++ src/main/scala/templemore/projecteuler/Problem14.scala
@@ -5,11 +5,11 @@ import scala.math.sqrt
 
 object Problem14 {
   def apply() = {
-    def stream(start: Int) = Stream.iterate(start)(i => if ( i % 2 == 0) i/2 else (3 * i) + 1)
+    def stream(start: Long) = Stream.iterate(start)(i => if ( i % 2L == 0) i/2L else (3L * i) + 1L)
 
-    ((1 to 999999).foldLeft(0, 0) { (acc, num) =>
+    ((1L to 999999L).foldLeft(0L, 0L) { (acc, num) =>
       val maxChainLength = acc._2
-      val len = (stream(num) takeWhile (_ > 1)).size
+      val len = (stream(num) takeWhile (_ > 1L)).size
       if ( len > maxChainLength ) (num, len) else acc
     })._1
   }
